requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
image==1.5.33
Markdown==3.3.4
Pillow==8.2.0
PyMySQL==1.0.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
